refactor: renamed: examples/listener/src/middlewares -> examples/listener/src/middleware
