+++ examples/listener/src/middleware/boomHandler.ts
